refactor: move votes and quorum stat to ProposalVotes component
